Add document download link and handler (#4011)
This PR adds a download link for each document on the documents
list. Clicking this link will open the document on a new tab or window.

To view these documents, a new endpoint is added which will fetch
and output the server's raw response. This required a refactor of the
API Client request method to handle responses that are not JSON,
and a new a property to return a raw response, instead of just the body.