Disable platform checkout if subscription product is in cart (#4020)
* Adding a utility class to check for platform_checkout availability

* Initialize platform checkout codes and script after cart is loaded from session

* Restore Platform_Checkout_Tracker after rebased

* Restore conditional checks for maybe_register_platform_checkout_hooks()

This function is now wrapped by Platform_Checkout_Utilities->should_enable_platform_checkout(),
thus the conditional check is redundant. But a few test cases test specifically against the
maybe_register_platform_checkout_hooks() method and the feature flags, I think it wouldn't hurt
to keep the conditional.

When we remove the feature flags all together, the tests can be removed/edited then.

* Add test for Platform_Checkout_Utilities

* Add changelog

* Use older phpunit method

onlyMethods is available in 9.5 but failed in pipeline. Use the older
function name which is setMethods.

* Only load the email field action and filter if there is no subscription product in cart

* Set false for config isPlatformCheckoutEnabled if subscription product in cart

Co-authored-by: David Stone <david@nnucomputerwhiz.com>