Tweak onboarding select control (#4054)
* Adjust LoadableBlock and move country description

* Update OnboardingSelectControl

Rename it to CustomSelectControl and make it usable anywhere by allowing the use of a custom component to render options and add a placeholder. Minor style changes.

* Update add business task

Replace OnboardingSelectControl with CustomSelectControl, adding placeholders and a custom option component for the business type. Minor style improvements to match the Figma design.

* Prevent custom select control button text overflow

* Optimize first businessCountry assignement

* Update test snapshots

* Sort countries by name rather than key