Add missing VAT information prompt when downloading invoice (#4140)
This PR adds a new hasSubmittedVatData flag to the wcpaySettings.accountStatus
object, based on the value we get from the basic account info call. This value is
then used when merchants click on the download button for a VAT invoice.

If the data has already been submitted, the document should be downloaded
right away. If the data hasn't been submitted, a modal should pop up with the
form to submit the VAT details.

We're only handling the modal as part of this PR. The actual form and data
submission will be handled in a separate PR for #3934.

Note: A few small fixes like removing unused props from the Document type,
removing a few extra break; statements, and re-adding the document data
methods to the wordpress/data store have also been included in this PR.