Add VAT form step to collect and validate the VAT number
This commit adds a step to the VAT form that collects and validates
a VAT number for the merchant. Merchants can specify that they don't
have a VAT number as well.

If the VAT number is invalid, an error message is displayed. If it is
valid, we attempt to grab the company name and address to reuse them
in subsequent steps.