Directly download documents from the documents list
This commit adds a handler to the documents list that will
automatically download a document, or go through the VAT
validation process, if the document ID and type are present
in the query string.

This is used to create a direct link from the new document
email to the actual report.