Add support for direct document access (#4180)
This PR modifies the document download button to use a helper method
and window.open instead of relying on the <a> href property to open the
document. This is useful to tidy up the retry mechanism and also to allow
documents to be downloaded without having a button on the page.

It also adds a handler to the documents list that will automatically download
a document, or go through the VAT validation process, if document_id and
document_type are present in the query string. This is used to create a
direct link from the new document email to the actual report.

Direct links will attempt to be opened in the current tab. However, depending
on the content-disposition header, files might be automatically downloaded.
In such cases, the current page will remain the documents list.