Add methods to manipulate live/test/dev mode for tests.
Those methods will triger a deprecation warning if used outside of unit tests (when `WCPAY_TEST_ENV` is not defined).

At some point/a future commit, we might decide to allow those methods to be used by consumers, and to remove filters from the `init` method. In that case, we should only allow such manipulations at an early stage (before a given hook for example).