Improve the existing locking mechanism
The following commits in this PR are going utilize the locking mechanism to more places in the code, and as discovered in the work for the next commits in this PR, it was not sufficient.

While testing concurrent processes, I realized that even though not worthless, the mechanism is not complete. Here's why:

If the order gets locked by one process before another one started, everything is fine. However, if both processes have already started, and especially on repetetive calls to `is_order_locked`, the same transient will be used, without removing it from the cache to force retrieval. After a bit of tinkering, this commit (as tested in the next ones), should bust all caching levels, and return a fresh value.