Upon API exceptions, wait for webhooks before fail
Imagine that the server successfully requested an intent creation+confirmation from Stripe,
but for any reason the response failed to make it back to the client. I wanted to simulate
that scenario in order to force webhooks to complete an order, and then (this will be in one
of the next commits) check that before creating a new intent. I used this snippet to simulate the error:

<?php
add_filter( 'wcpay_api_request_response', 'fail_intent_response', 10, 4 );
function fail_intent_response( $response, $method, $url, $api ) {
	if ( 'POST' === $method && 'intentions' === $api ) {
		$response['body'] = '\\invalid_json';
	}
	return $response;
}
?>

While testing with it, I saw the following happen:

1. Order is created (before we step in)
2. A `create_and_confirm_intent` request is sent to the server, and forwarded to Stripe.
2. Webhook marks the order as paid.
3. __Gateway marks the order as failed.__
4. Removing the snippet, and re-submitting the form causes a new intent to be created, resulting in the following order notes:
	1. Payment successful with intent A
	2. Payment failed
	3. Payment successful with intent B

In this case intent B's details overwrite intent A's details, meaning that not only are there two
intents, but intent A is no longer associated with an order, and cannot be refunded anymore.

To fix this, based on the improved locking mechanism in the previous commit in this branch, the following is done in this commit:

- A new `is_order_paid_regardless_of_exception` method is introduced in the order service.
- The new method is the first thing that is checked in the gateway upon an exception. If the method returns true, then the gateway returns a successful redirect, if not proceeds as usual.

`is_order_paid_regardless_of_exception()` has thorough comments, but overall it does the following:

1. Check if the exception is an API_Exception. Others do not matter here.
2. Wait for 5 (kind of arbitrary) seconds for the webhook to kick-in.
3. Check if the order is locked, and if it is, retry each second until the order gets unlocked, up to 15 times/seconds. This will allow the webhook-triggered status change to complete all necessary actions, including 3rd-party ones.
4. Only then, once the order is unlocked, check if it has a paid status now.

This way, even if there is such an exception the order notes will look like this:

1. Success
(that's it, no failure here)

At the same time, the customer will be redirected straight to the thank-you page, not knowing that there was anything wrong, and not having had the opportunity to re-submit the form.

Note: This only applies to classic checkout.