Fix error sorting the Uncaptured table using Capture by column (#5185)
This change fixes an error in the Uncaptured transactions table, when the table is sorted clicking on the Capture by column.
The error is caused by a 500 response from the endpoint, because 'capture_by' is not a valid sorting field in the API, since it is a value calculated in the front end, derived from the 'created' field.

To solve the issue, I am rewriting the orderby field in the resolver to 'created', in case the 'capture_by' field is specified. This has the same effect, and does not cause an error in the API endpoint.