Initial implementation.
Right now there is an order payment class, but I do not like how it is loaded. Ideally there should be an order payment class, but it should not manage the link between itself and the order. In the next iteration, that will be separate to support a scenario where an order could potentially have multiple payments.