Improve the logic which determines if a subscriptions plugin is being activated
Inside the function to check if we need to load subscriptions-core functionality
we have a condition that checks if the subscriptions plugin is in the process
of being activated, the purpose of this check is to avoid fatal errors due to load
orders.

This commit tightens the logic to avoid running these checks on all page loads.