Use the billing_cycle_anchor to set a free trial
We need to use the billing_cycle_anchor instead of the trial_end
because Stripe will automatically start the subscription if it has a
free trial before we've had a chance collect a payment method.

This can lead to multiple Subscriptions being created at Stripe
if payment collection fails on checkout.

To ensure there's a way for us to recognise subscription create requests
that contain free trial, I've updated the line item metadata we send to
stripe to make sure we flag items with free trials.