Prevent multiple iFrames from rendering
Before, the getProductData() and addToCart() functions were causing the
newIframe() callback to be recreated, ultimately causing the useEffect()
that adds the iFrame into the DOM to add another iFrame. Now, the use of
useRef() guarantees we have access to getProductData() and addToCart()
but without causing the newIframe() callback to recreate the iFrame.