Fix: WooPay button from enabling before redirecting to WooPay checkout
This occurs in the product page.

The reason it occurred, was because an `addToCart()` HTTP request is
made, which triggers the `updated_cart_totals` event to fire. This event
then causes a new WooPay button to be rendered. The new button is
enabled by default and so it appears as if the WooPay button gets
enabled right before the browser redirects to the WooPay checkout page.