Reinitialize the paymentRequest if the customer previously aborted a request
This ensures that the available shipping methods are fetched again as they
may be different to what was previously proviced when the customer clicked
the express payment button.

Also reinit the payment button if the product's shipping status is different
to the last products shipping status. ie virtual -> physical -> virtual.
That flow needs to reinit the form because the needs shipping payment button param
is not editable/updatable. A new request must be created.