Avoid setting tk_ai cookie on page load
tracksUserIdentity is printed onto the page via payment_fields when the
page loads. This is making the page uncacheable. With this change we
prevent the setting of the cookie on the page load. Instead the frontend
will try to use the Tracks identity from the cookie if it's set,
otherwise it will retrieve it via an Ajax requests when needed.