Fix WCPay Tracks events recorded from Admin area
Tracks events recorded from Admin area are missing the identity string.
This is a side effect of a recent additional check, which is being
removed with this commit.
This check is redundant as we check the Tracking conditions before
calling the tracks_get_identity function, where we can pass the
additional parameters to flag events from the admin area.