Prefetch WooPay session data on page load
Previously, the WooPay session data would be retrieved when the shopper
clicked on the "Proceed to Checkout" button. Now, it's prefetched during
page load. So that when the shopper clicks the "Proceed to Checkout"
button, it won't have to also make this request. Therefore, making the
Direct Checkout flow faster.