Fix: CLASSIC_CART_PROCEED_BUTTON should reference '.wc-proceed-to-checkout .checkout-button'
Before, the CLASSIC_CART_PROCEED_BUTTON was referencing the
"Proceed to Checkout" button's parent, in the classic cart. This was
unintentionally causing the click handler to be applied on the "Proceed
to Checkout" and "Buy with WooPay" buttons. Now, it correctly references
the "Proceed to Checkout" button. So, now the click handler is being
applied correctly.