Use the generated_card for subscriptions
`setup_future_use` means that the returned payment intent has an additional payment method, associated via the charge object. This is under `generated_card`.

The `generated_card` is a `card` payment method, not `card_present`, and it's associated with the customer.

Storing this one allows us to properly charge the customer outside of the initial transaction.