Ensure a timeout fail-safe exists when a message is posted to WooPay
Previously, if a message was posted to WooPay and WooPay did not
respond, it would be possible for the DC flow to hang. Now, a timeout
is added so that if a response is not returned within the allotted time,
an error is thrown and the DC flow short-circuits back to the standard
checkout flow.