Revert if statement on settings REST controller
In order to avoid masking or silencing this and other failures, we're reverting the if statement added previously. Since no issues were reported so far, most likely this won't make a big difference.

Plus, there are other calls to  in the same file without extra checks, we should be on the safe side with this removal.